requirements.txt
streamlit>=1.29.0
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.1
spacy>=3.7.0
sqlalchemy>=2.0.0
pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
tqdm>=4.65.0
python-dateutil>=2.8.2
